revert(types): keep PERMISSIONS in packages/api — the move needs a /publish first (#5194)
Deploy Validation, a REQUIRED check, caught this. Nothing local could.

#5191 item 3's premise was: *"Both consumers resolve it as `workspace:*`, so no
npm publish is required to land this."* True of the monorepo, FALSE of the
scaffold lane — and that is the lane that gates the merge.

`create-atlas/templates/{docker,nextjs-standalone}` depend on the PUBLISHED
`@useatlas/types` (`^0.7.0` → 0.10.0 on npm), and Deploy Validation builds
`packages/api` against that copy. So the moment this module re-exported the
runtime tuple from `@useatlas/types`:

    Error: Turbopack build failed with 7 errors:
    Error: Export PERMISSIONS doesn't exist in target module

Verified directly: `npm pack @useatlas/types@0.10.0` ships `ATLAS_ROLES` in
`dist/auth.js` and no `PERMISSIONS`.

This is CLAUDE.md's "sequence ref bumps AFTER the publish lands" rule arriving
through an unwatched door — it is usually about scaffolds failing on `npm
install`, and here it is a bundler resolving a missing named export.

⚠️ The local pre-flight structurally CANNOT see it: `bun run type`, `bun run
lint` and both sweeps pass, because the monorepo resolves `workspace:*`. Deploy
Validation is the only job that builds against npm. Same shape as the
"Symlink Stub Build" finding two commits ago — a second CI job proving a
property the local tree cannot.

Split per the maintainer's call: the #5192 security fix is time-sensitive (it
must land before the next `/release`), so it ships now and the promotion becomes
#5194, sequenced after a `/publish` of `@useatlas/types`.

## What reverts

- `PERMISSIONS` / `Permission` / `isValidPermission` return to
  `packages/api/src/lib/auth/permissions.ts`, with the full retraction in the
  docstring so the next attempt does not rediscover it.
- `packages/types/src/auth.ts` loses the block — and with it the decision to
  put a permission tuple on the public npm surface, which the type reviewer had
  flagged as needing an explicit answer. #5194 asks it.
- `PERMISSION_LABELS` goes back to `Record<string, string>`, with the missing
  exhaustiveness named as a KNOWN GAP rather than left silent.
- The web test goes back to a hand-copy — honestly labelled, pointing at #5194.

## What SURVIVES, because none of it needed the move

- `dashboards:share` in the label map and the Dashboards group (#5192).
- `permissionLabel` going through a `ReadonlyMap` — the prototype hole where
  `"toString"` returned a **Function** past a `?? p` fallback, from a function
  declaring `: string`. Needs no shared type.
- Its runtime test, including the prototype keys, and the "every known flag has
  real copy" assertion that is now the only thing holding the label map honest.
- The seven API mocks spreading the REAL tuple and using the REAL
  `isValidPermission` — repointed at `@atlas/api/lib/auth/permissions`, which is
  where it lives again. That was the largest single defect in item 3 (all seven
  copies were stale) and it does not regress.

Verified by reproducing BOTH CI jobs locally: the stub build
(`rm -rf ee && cp -R scripts/ee-stub ee && bun install && bun run type`) and a
`npm pack` of the published types package. `ee/` restored from a copy, not
`git checkout`.

api --affected 324/324 · web 335/335 · -pg 10/10 · lint + type clean

Refs #5191, #5194
